requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
CacheControl==0.13.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
drf-yasg==1.21.6
future==0.18.3
geocoder==1.38.1
geojson==2.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
install==1.3.5
joblib==1.3.1
jsonpickle==0.9.6
msgpack==1.0.5
nltk==3.8.1
packaging==23.1
phonenumbers==8.13.16
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
weather-api==1.0
